hosting-automator/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.2
python-dotenv==1.0.0

# Database
supabase==2.0.2

# SSH and CLI automation
paramiko==3.4.0

# HTTP requests for Matomo API
requests==2.31.0

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0